Revert "feat(data-deletion): refactor Person bulk_delete and make code usable… (#56587)"
This reverts commit be1b98a34567c7ca9be03aef2d8476133397f6c2.
